webserver/requirements.txt
flask-cors
numpy==1.16.5
diskcache
flask
flask_restful
futures==3.1.1
pymilvus==0.3.0
PubChemPy
psycopg2-binary